Move listener class, remove visibleForTesting isInitialized.
